BPTreeIndexTest does consistency check in the end

+++ community/index/src/test/java/org/neo4j/index/bptree/IndexModifierTest.java
@@ -103,7 +103,7 @@ public class IndexModifierTest
         for ( int i = 0; i < maxKeyCount; i++ )
         {
             // given
-            insert( (long) i, (long) i );
+            insert( i, i );
 
             // then
             assertThat( keyAt( i ), is( (long) i ) );
